refactor: contract
